Fix eval_utils.display_group()
At line 44, the value of a foreground pixel starts from 0. In this way, pixels of the first group from reg_groups will be all 0 and will blend in the background since values of background pixels are all 0 too. 
Change to start the value to give from 1 so all foreground pixels have different values from background pixels.